backend/requirements.txt
asgiref==3.7.2
auditlog==0.0.5
Babel==2.12.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-jsonfield==1.4.1
django-money==3.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
packaging==23.2
Pillow==10.0.0
protobuf==4.24.3
psycopg2-binary==2.9.6
py-moneyed==3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymemcache==4.0.0
pyparsing==3.1.1
python-dateutil==2.6.0
python-memcached==1.59
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.6.0
